+++ app/src/main/java/com/codesoom/assignment/controllers/TaskController.java
@@ -1,12 +1,10 @@
 package com.codesoom.assignment.controllers;
 
-import com.codesoom.assignment.TaskIdNotFoundException;
+import com.codesoom.assignment.TaskService;
 import com.codesoom.assignment.models.Task;
-import com.codesoom.assignment.repository.TaskRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
 
@@ -15,50 +13,39 @@ import java.util.List;
 @CrossOrigin
 public class TaskController {
 
-    TaskRepository taskRepository = new TaskRepository();
+    private final TaskService taskService;
+
+    public TaskController(TaskService taskService) {
+        this.taskService = taskService;
+    }
 
     @GetMapping
-    public List<Task> getTaskList() {
-        return new ArrayList<>(taskRepository.getTasks());
+    public List<Task> handleGetTasks() {
+        return taskService.findAllTasks();
     }
 
     @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public Task getTaskById(@PathVariable Long id) {
-        return findTask(id);
+        return taskService.findTask(id);
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public Task handleCreate(@RequestBody Task task) {
-        task.setId(taskRepository.generateId());
-        taskRepository.createTask(task.getId(), task);
-        return task;
+        return taskService.createNewTask(task);
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public Task handleUpdate(@PathVariable Long id, @RequestBody Task task) {
-        findTask(id);
-        Task editTask = task;
-        editTask.setId(id);
-        taskRepository.updateTask(task.getId(), editTask);
-        return task;
+        return taskService.editTask(id, task);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void handleDelete(@PathVariable Long id) {
-        findTask(id);
-        Task task = findTask(id);
-        taskRepository.deleteTask(id);
-    }
-
-    public Task findTask(Long id) {
-        if (taskRepository.findTaskWithIdInTasks(id) == null) {
-            throw new TaskIdNotFoundException();
-        }
-        return taskRepository.findTaskWithIdInTasks(id);
+        taskService.removeTask(id);
     }
 
 }
